added videos to participants
